[ISSUE4] CommonExceptionHandler 사용
